requirements.txt
httpx==0.27.2
selectolax==0.3.21
python-dotenv==1.0.1
duckdb==1.1.1
